+++ CHANGELOG.md
@@ -1,3 +1,4 @@
-[1.0.0]
+## [1.0.0 - 03/09/2023]
 
--
+- Implement cookies for login [#1](https://github.com/BIEMAX/quote-manager/issues/1)
+- Implement navigation guard [#2](https://github.com/BIEMAX/quote-manager/issues/2)